+++ src/sources/SourceDataMemory.cpp
@@ -1,6 +1,6 @@
 #include "SourceDataMemory.hpp"
 
-SourceDataMemory::SourceDataMemory(uint8_t min, uint8_t max) : SourceData(min, max)
+SourceDataMemory::SourceDataMemory(uint64_t min, uint64_t max) : SourceData(min, max)
 {
 }
