+++ src/sources/SourceDataNetworkUsedBandwith.cpp
@@ -1,6 +1,6 @@
 #include "SourceDataNetworkUsedBandwith.hpp"
 
-SourceDataNetworkUsedBandwith::SourceDataNetworkUsedBandwith(uint8_t min, uint8_t max) : SourceData(min, max)
+SourceDataNetworkUsedBandwith::SourceDataNetworkUsedBandwith(uint64_t min, uint64_t max) : SourceData(min, max)
 {
 }
